+++ Assets/Scripts/FirstPersonController.cs
@@ -265,7 +265,7 @@ public class FirstPersonController : MonoBehaviour
             ball.TryKick(gameObject);
     }
 
-    private const string GameVersion = "0.3.29-dev";
+    private const string GameVersion = "0.3.30-dev";
 
     private float lastSpeed;
     private bool lastSprinting;